// src/shared/models/ogame/messages/OgameRawExpeditionResultType.ts
export enum OgameRawExpeditionResultType {
    resources = 'ressources',
    nothing = 'nothing',
    fleet = 'shipwrecks',
    combatPirates = 'combatPirates',
    combatAliens = 'combatAliens',
    item = 'items',
    delayOrEarly = 'navigation',
    lostFleet = 'fleetLost',
    darkMatter = 'darkmatter',
    trader = 'trader',
}